added custom hooks for auth actions
